add size_t needed header

+++ common/reflectivity/reflectivity.cpp
@@ -8,11 +8,11 @@
 using namespace rs2;
 
 
-static const int   N_STD_FRAMES = 100;
-static const int   NINETY_FIVE_PERCENT_OF_STD_PERIOD = static_cast< int >( 0.95 * N_STD_FRAMES );
-static const float MAX_RANGE_IN_UNIT = 65536.0f;
-static const float LONG_THERMAL = 74.5f;
-static const float INDOOR_MAX_RANGE = 9.0f;
+static const size_t N_STD_FRAMES = 100;
+static const int    NINETY_FIVE_PERCENT_OF_STD_PERIOD = static_cast< int >( 0.95 * N_STD_FRAMES );
+static const float  MAX_RANGE_IN_UNIT = 65536.0f;
+static const float  LONG_THERMAL = 74.5f;
+static const float  INDOOR_MAX_RANGE = 9.0f;
 
 // TODO try to read from LRS
 static const float FOV_H = 0.610865f;
